new(eslint-config): Migrate to presets and disable some rules.
Disabled rules:
- node/no-unpublished-import
- unicorn/prefer-set-has
- unicorn/explicit-length-check
- @typescript-eslint/prefer-regexp-exec
- @typescript-eslint/restrict-template-expressions
- promise/catch-or-return

Updated rules:
- complexity bumped to 20
