+++ packages/config-eslint/src/presets/next.ts
@@ -1,0 +1,13 @@
+import { ESLintConfig } from '@beemo/driver-eslint';
+
+import unicornRules from '../rules/unicorn';
+
+const config: ESLintConfig = {
+  extends: ['plugin:eslint-plugin-unicorn/recommended'],
+  plugins: ['unicorn'],
+  rules: {
+    ...unicornRules,
+  },
+};
+
+export = config;
