+++ packages/config-eslint/src/presets/node.ts
@@ -1,0 +1,20 @@
+import { ESLintConfig } from '@beemo/driver-eslint';
+
+import nodeRules from '../rules/node';
+
+const config: ESLintConfig = {
+  env: {
+    browser: false,
+    node: true,
+  },
+  settings: {
+    node: {
+      tryExtensions: ['.js', '.jsx', '.json'],
+    },
+  },
+  rules: {
+    ...nodeRules,
+  },
+};
+
+export = config;
