+++ packages/config-eslint/src/rules/prettier.ts
@@ -1,12 +1,8 @@
 import { ESLintConfig } from '@beemo/driver-eslint';
 
-const config: ESLintConfig = {
-  extends: ['eslint-config-prettier'],
-  plugins: ['eslint-plugin-prettier'],
-  rules: {
-    // eslint-plugin-prettier rules
-    'prettier/prettier': 'error',
-  },
+const config: ESLintConfig['rules'] = {
+  // eslint-plugin-prettier rules
+  'prettier/prettier': 'error',
 };
 
 export = config;
